Avoid test name collision
